Prevent from crash in expo-sensors when quitting react activity multiple times.

+++ packages/expo-sensors/android/src/main/java/expo/modules/sensors/services/SensorServiceSubscription.java
@@ -3,6 +3,7 @@
 package expo.modules.sensors.services;
 
 import android.hardware.SensorEventListener2;
+import android.util.Log;
 
 public class SensorServiceSubscription implements org.unimodules.interfaces.sensors.SensorServiceSubscription {
   private boolean mIsEnabled = false;
@@ -20,6 +21,7 @@ public class SensorServiceSubscription implements org.unimodules.interfaces.sens
     assertSubscriptionIsAlive();
     if (!mIsEnabled) {
       mIsEnabled = true;
+      Log.d("SensorServiceSub", "start: " + this);
       mSubscribableSensorService.onSubscriptionEnabledChanged(this);
     }
   }
@@ -42,7 +44,6 @@ public class SensorServiceSubscription implements org.unimodules.interfaces.sens
   }
 
   public void stop() {
-    assertSubscriptionIsAlive();
     if (mIsEnabled) {
       mIsEnabled = false;
       mSubscribableSensorService.onSubscriptionEnabledChanged(this);
@@ -50,9 +51,10 @@ public class SensorServiceSubscription implements org.unimodules.interfaces.sens
   }
 
   public void release() {
-    assertSubscriptionIsAlive();
-    mSubscribableSensorService.removeSubscription(this);
-    mHasBeenReleased = true;
+    if(!mHasBeenReleased) {
+      mSubscribableSensorService.removeSubscription(this);
+      mHasBeenReleased = true;
+    }
   }
 
   private void assertSubscriptionIsAlive() {
